+++ tests/testapp/settings.py
@@ -33,7 +33,7 @@ SECRET_KEY = "supersikret"
 LOGIN_REDIRECT_URL = "/?login=1"
 
 ROOT_URLCONF = "testapp.urls"
-LANGUAGES = (("en", "English"), ("de", "German"))
+LANGUAGES = [("en", "English"), ("de", "German"), ("fr", "French")]
 
 TEMPLATES = [
     {